requirements.txt
requests
beautifulsoup4
pysocks
colorama
watchdog
stem
pyngrok
fake-useragent
